Finished with filtering posts based on tags.
Functions created
filteringPosts()  with the argument of filteredPosts  to loop through the array of objects generate the html content to populate the container.
filterByCategory which takes as argument the tags categories filters the array of object and return the ones that have the arguments category.
and last the filtering post function is called again and populates the container